Refactored GameSession, added Name
