+++ api/api/public/gamesession/models.py
@@ -1,6 +1,5 @@
 from typing import List
 
-import uuid as uuid_pkg
 from sqlmodel import Field, Relationship, SQLModel
 
 from api.public.room.models import Room
@@ -8,15 +7,14 @@ from api.public.player.models import Player
 
 
 class GameSessionBase(SQLModel):
-    pass
+    name: str
+    author: str
 
 
 class GameSession(GameSessionBase, table=True):
     id: int = Field(default=None, primary_key=True)
-    uuid: str = Field(
-        default_factory=lambda: str(uuid_pkg.uuid4()),
-        nullable=False,
-    )
+    name: str = Field(max_length=30, unique=True)
+    author: str = Field(max_length=50)
 
     rooms: List["Room"] = Relationship(back_populates="gamesession")
     players: List["Player"] = Relationship(back_populates="gamesession")
@@ -27,8 +25,9 @@ class GameSessionCreate(GameSessionBase):
 
 
 class GameSessionRead(GameSessionBase):
-    uuid: str | None = None
+    id: int
 
 
 class GameSessionUpdate(GameSessionBase):
-    uuid: str | None = None
+    name: None | str = None
+    author: None | str = None
